split out right click calls
