+++ OPmenu.xml
@@ -49,25 +49,32 @@ return "Unavailable in this Context"
 n = kwargs.get("node", None)
 if n.type().name() == "redshift_vopnet":
     import matlib
-    matlib.main_panel.save_material(n)
+    from matlib.utils import rc_calls
+    rc_calls.save_material(n)
 elif n.type().name() == "octane_vopnet":
     import matlib
-    matlib.main_panel.save_material(n)
+    from matlib.utils import rc_calls
+    rc_calls.save_material(n)
 elif n.type().name() == "materialbuilder":
     import matlib
-    matlib.main_panel.save_material(n)
+    from matlib.utils import rc_calls
+    rc_calls.save_material(n)
 elif n.type().name() == "principledshader::2.0":
     import matlib
-    matlib.main_panel.save_material(n)
+    from matlib.utils import rc_calls
+    rc_calls.save_material(n)
 elif n.type().name() == "arnold_materialbuilder":
     import matlib
-    matlib.main_panel.save_material(n)
+    from matlib.utils import rc_calls
+    rc_calls.save_material(n)
 elif n.type().name() == "subnet":
     import matlib
-    matlib.main_panel.save_material(n)
+    from matlib.utils import rc_calls
+    rc_calls.save_material(n)
 elif n.type().name() == "collect":
     import matlib
-    matlib.main_panel.save_material(n)
+    from matlib.utils import rc_calls
+    rc_calls.save_material(n)
 else:
     hou.ui.displayMessage('Please select Material Builder Nodes to save Material')
